W8-O-5d-Q1/Q2: Zahlenrand an den Betriebsschwellen, keine int-Abschneidung mehr im BHKW-Plan-Port
Zwei Anwenderentscheide vom 07.09.2026 ("Empfehlung" bzw. "keine Treue zur alten
DLL") loesen die drei Schwellen ab, an denen die Umstellung auf double eine
Differenz im letzten Bit in eine ganze Fahrweise uebersetzte (Befund im
protokoll.txt der Basis 2026-09-07_R4_Double: elf der zwoelf Referenzprojekte).

Q1 - EIN benannter Zahlenrand, neu in EPOS.Kern/Allgemein/Simulation/Rechenrand.cs:

    Rand = ABSOLUT (1e-9) + RELATIV (1e-12) * |Schwelle|
    Rechenrand.SchwelleErreicht(wert, schwelle)  =>  wert >= schwelle - Rand

Absolut UND relativ, weil die Schwellen Energien in kWh tragen und mehrere
Groessenordnungen spannen: ein rein absoluter Rand waere bei 100 000 kWh zu knapp
(ein ulp misst dort 1,5e-11 und die Schwelle entsteht aus mehreren Rechnungen),
ein rein relativer verschwaende an einer Schwelle von 0. Der Rand bleibt vier
Groessenordnungen unter der schaerfsten Vergleichstoleranz der Referenzsuite
(rel. 1e-4) - er kann keine Abweichung erzeugen, die ein Referenzvergleich saehe.

Angewandt an den zwei Schwellen des Befunds, die alte Bauart je im Kommentar:

  * SimulationPufferspeicher.HystereseFortschreiben - "SOC >= Q_max * SchwelleAus"
    gegen einen Stand, den Ladefaehigkeit auf GENAU Q_max * grenze faehrt; in
    Gleitkomma ist a + (b - a) nicht bitgleich b. Bistabil, der Fehltritt trug
    ueber Stunden weiter. Die EINschaltschwelle bleibt bewusst ohne Rand: Auf sie
    steuert kein Rechenweg den Fuellstand an.
  * SimulationBHKW.Motorlauf_Waermegefuehrt, beide Stufen - Volllast gegen
    Modulation und Modulation gegen "Motor aus". Der Fall Gleichheit wechselt bei
    der ersten Stufe von der Modulation in die Volllast; beide Zweige rechnen dort
    dasselbe Ergebnis, und BHKW.wiki nennt die Volllast ohnehin fuer P_th <= R_Q+S.

Q2 - die drei Physik-Funktionen in BhkwPlan.cs geben double zurueck; die (int)-
Abschneidung (Borland _ftol) faellt:

  * TaeglHeizlastWG, SolareGewinneC, SpezWaermeverlusteC
  * Die Aufrufer in SimulationWaermebedarf folgen mit. Hinter
    SpezWaermeverlusteC stand "/ 100" auf einem int - eine GANZZAHLIGE Division,
    also ein zweites Abschneiden hinter dem ersten. Sie heisst jetzt "/ 100.0".
    Der Faktor 100 selbst bleibt: Er gehoert zur Schnittstelle, nicht zur Physik.

Nachweise, je mit Gegenprobe:

  * EPOS.Kern.Tests/RechenrandTests (6 Faelle) - je Schwelle ein Stand GENAU auf
    der Grenze und einer ein ulp darunter (Math.BitDecrement); beide liefern
    dieselbe Entscheidung, und der blanke Vergleich auf denselben zwei Zahlen
    faellt verschieden aus. Dazu die Bistabilitaetsprobe: fuellen ueber
    Ladefaehigkeit/Laden auf Q_max * grenze, dann sechs Stunden stabil - und die
    Hysterese lebt weiter, unter der Einschaltschwelle springt sie neu an.
  * EPOS.Kern.Tests/BhkwPlanRueckgabeTests (5 Faelle) - die Nachkommastellen
    ueberleben, die Flaechenskalierung wirkt ungerundet (zwei Gebaeude, die frueher
    dieselbe ganze Zahl lieferten), und ein Waechter haelt den Rueckgabetyp.

Hausregeln in EPOS.Kern/CLAUDE.md: zwei neue Abschnitte - "Vergleiche an
Betriebsschwellen tragen den Zahlenrand" (wer eine neue Schwelle einfuehrt, nimmt
SchwelleErreicht, nicht >=) und "Keine (int)-Abschneidung auf einer Rechengroesse".

Wiki nachgezogen, ohne neue Formelnummer (die Waechter bleiben gruen):
Pufferspeicher.wiki (Rand an (8), warum (7) ihn nicht braucht), BHKW.wiki (beide
Grenzen der waermegefuehrten Tabelle) und Waermebedarf.wiki ((2), (7) und (12)
werden ungerundet weitergereicht - das Raster ist ersatzlos entfallen).

Der Referenzlauf aendert sich dadurch; die neue Basis 2026-09-07_R5_Zahlenrand
kommt im dritten Commit dieser Reihe.

Kern-Tests 1 839 (vorher 1 828), UI 3 126, SpeicherEngine 337, KiKern 469,
Formularkarte 122, ChartProben 44 Bilder alle gruen, SQL-Pruefer 0 Fundstellen,
ResourceDesigner abweichend 0, WP-Plan.sln Release x64 ohne neue Warnung.

+++ EPOS.Kern/Allgemein/Simulation/SimulationWaermebedarf.cs
@@ -732,6 +732,19 @@ namespace WindowsFormsApplication1
                 }
             }
 
+            // ANWENDERENTSCHEID W8-O-5d-Q2 (07.09.2026): "keine Treue zur alten DLL".
+            // Die drei Physik-Funktionen des BHKW-Plan-Ports gaben bis hierher int zurueck
+            // (Borland _ftol, Abschneiden Richtung Null); seither geben sie double zurueck.
+            // Zwei Stellen ziehen damit mit:
+            //
+            //   * Die Division "/ 100" hinter SpezWaermeverlusteC war GANZZAHLIG, solange
+            //     die Funktion int lieferte - also ein zweites Abschneiden hinter dem
+            //     ersten. Sie steht jetzt als "/ 100.0" da und ist eine double-Division.
+            //   * Die Division hinter SolareGewinneC war schon immer eine double-Division;
+            //     "(double)100" ist nur noch "100.0" geschrieben - derselbe Wert.
+            //
+            // Der Faktor 100 selbst BLEIBT: Er gehoert zur Schnittstelle der DLL-Funktion
+            // (sie liefert das Hundertfache), nicht zur Physik.
             for (int Tag = 350; Tag < 365; Tag++)
             {
                 /*
@@ -741,7 +754,7 @@ namespace WindowsFormsApplication1
                 */
                 Solare_Gewinne[Tag] = WPPlan.Core.BhkwPlan.SolareGewinneC(Sol_N[Tag], (double)item.Fensterflaeche_Nord, Sol_w[Tag], Sol_O[Tag],
                         (double)item.Fensterflaeche_Ost, Sol_S[Tag], (double)item.Fensterflaeche_Sued,
-                        (double)item.Fensterdurchlassgrad) / (double)100;
+                        (double)item.Fensterdurchlassgrad) / 100.0;
 
                 /*
                 SpezWaermeverluste[Tag] = com.I_SpezWaermeverlusteC((double)item.k_Wert_Außenwand, (double)item.Flaeche_Außenwand,
@@ -758,7 +771,7 @@ namespace WindowsFormsApplication1
                        (double)item.k_Wert_Sonstiges, (double)item.Sonstige_Flaechen, (double)item.Waermebrueckenverlustkoeffizient_Anschluß_Fenster_Wand,
                        (double)item.Abmessung_Anschluß_Fenster_Wand, (double)item.Waermebrueckenverlustkoeffizient_Anschluß_Wand_Dach, (double)item.Abmessung_Anschluß_Wand_Dach,
                        (double)item.Waermebruckenverlustkoeffizient_Anschluß_Außenwand_Kellerdecke, (double)item.Abmessung_Anschluß_Außenwand_Kellerdecke, A_Temp[Tag], (double)item.Wohnflaeche,
-                       (double)item.Raumhoehe, (double)item.Luftwechselrate) / 100;
+                       (double)item.Raumhoehe, (double)item.Luftwechselrate) / 100.0;
 
                 WE_Absenkung = 0;
                 if ((double)item.Raumsolltemperatur_Wochenende > 5)
@@ -811,7 +824,7 @@ namespace WindowsFormsApplication1
                 */
                 Solare_Gewinne[Tag] = WPPlan.Core.BhkwPlan.SolareGewinneC(Sol_N[Tag], (double)item.Fensterflaeche_Nord, Sol_w[Tag], Sol_O[Tag],
                     (double)item.Fensterflaeche_Ost, Sol_S[Tag], (double)item.Fensterflaeche_Sued,
-                    (double)item.Fensterdurchlassgrad) / (double)100;
+                    (double)item.Fensterdurchlassgrad) / 100.0;
                 /*
                 SpezWaermeverluste[Tag] = com.I_SpezWaermeverlusteC((double)item.k_Wert_Außenwand, (double)item.Flaeche_Außenwand,
                         (double)item.k_Wert_Fenster, (double)item.gesamte_Fensterflaeche, (double)item.k_Wert_Dachflaeche,
@@ -827,7 +840,7 @@ namespace WindowsFormsApplication1
                      (double)item.k_Wert_Sonstiges, (double)item.Sonstige_Flaechen, (double)item.Waermebrueckenverlustkoeffizient_Anschluß_Fenster_Wand,
                      (double)item.Abmessung_Anschluß_Fenster_Wand, (double)item.Waermebrueckenverlustkoeffizient_Anschluß_Wand_Dach, (double)item.Abmessung_Anschluß_Wand_Dach,
                      (double)item.Waermebruckenverlustkoeffizient_Anschluß_Außenwand_Kellerdecke, (double)item.Abmessung_Anschluß_Außenwand_Kellerdecke, A_Temp[Tag], (double)item.Wohnflaeche,
-                     (double)item.Raumhoehe, (double)item.Luftwechselrate) / 100;
+                     (double)item.Raumhoehe, (double)item.Luftwechselrate) / 100.0;
 
                 WE_Absenkung = 0;
                 if ((double)item.Raumsolltemperatur_Wochenende > 5)
